+++ src/main/java/amlsim/model/aml/StackTypology.java
@@ -24,7 +24,7 @@ public class StackTypology extends AMLTypology {
 //        return orig_members * mid_members + mid_members + bene_members;
 //    }
 
-    StackTypology(float minAmount, float maxAmount, int minStep, int maxStep) {
+    StackTypology(double minAmount, double maxAmount, int minStep, int maxStep) {
         super(minAmount, maxAmount, minStep, maxStep);
     }
 
@@ -41,9 +41,9 @@ public class StackTypology extends AMLTypology {
         int mid_members = orig_members;  // Second 1/3 accounts are intermediate accounts
         int bene_members = total_members - orig_members * 2;  // Rest of accounts are beneficiary accounts
 
-        float amount1 = getRandomAmount();
-        float total_flow = amount1 * orig_members * mid_members;  // Total transaction amount
-        float amount2 = total_flow / (mid_members * bene_members);
+        double amount1 = getRandomAmount();
+        double total_flow = amount1 * orig_members * mid_members;  // Total transaction amount
+        double amount2 = total_flow / (mid_members * bene_members);
 
         for(int i=0; i<orig_members; i++){  // originator accounts --> Intermediate accounts
             Account orig = alert.getMembers().get(i);
